Project reorganization and cleanup
