+++ app/build.gradle
@@ -2,13 +2,13 @@ apply plugin: 'com.android.application'
 
 project.ext.ASSET_DIR = projectDir.toString() + '/src/main/assets'
 
-assert file(project.ext.ASSET_DIR + "/model_2.tflite").exists()
-assert file(project.ext.ASSET_DIR + "/labels_2.txt").exists()
+assert file(project.ext.ASSET_DIR + "/model.tflite").exists()
+assert file(project.ext.ASSET_DIR + "/labels.txt").exists()
 
 android {
     compileSdkVersion 28
     defaultConfig {
-        applicationId "org.tensorflow.lite.examples.classification"
+        applicationId "flower.classifier"
         minSdkVersion 21
         targetSdkVersion 28
         versionCode 1
@@ -20,7 +20,7 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
-    // TODO: Add an option to avoid compressing TF Lite model file
+
     aaptOptions {
         noCompress "tflite"
     }
@@ -50,10 +50,7 @@ dependencies {
     // CameraX View class
     implementation 'androidx.camera:camera-view:1.0.0-alpha19'
     // Build off of nightly TensorFlow Lite
-    // TODO: Add TFLite dependencies
     implementation('org.tensorflow:tensorflow-lite:0.0.0-nightly') { changing = true }
     implementation('org.tensorflow:tensorflow-lite-gpu:0.0.0-nightly') { changing = true }
     implementation('org.tensorflow:tensorflow-lite-support:0.0.0-nightly') { changing = true }
-    // Use local TensorFlow library
-    // implementation 'org.tensorflow:tensorflow-lite-local:0.0.0'
 }
