+++ app/src/main/java/flower/classifier/LegacyCameraConnectionFragment.java
@@ -1,4 +1,4 @@
-package org.tensorflow.lite.examples.classification;
+package flower.classifier;
 
 /*
  * Copyright 2019 The TensorFlow Authors. All Rights Reserved.
@@ -31,11 +31,13 @@ import android.view.Surface;
 import android.view.TextureView;
 import android.view.View;
 import android.view.ViewGroup;
+
+
 import java.io.IOException;
 import java.util.List;
-import org.tensorflow.lite.examples.classification.customview.AutoFitTextureView;
-import org.tensorflow.lite.examples.classification.env.ImageUtils;
-import org.tensorflow.lite.examples.classification.env.Logger;
+import flower.classifier.customview.AutoFitTextureView;
+import flower.classifier.env.ImageUtils;
+import flower.classifier.env.Logger;
 
 public class LegacyCameraConnectionFragment extends Fragment {
   private static final Logger LOGGER = new Logger();
