+++ app/src/main/java/flower/classifier/customview/AutoFitTextureView.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.tensorflow.lite.examples.classification.customview;
+package flower.classifier.customview;
 
 import android.content.Context;
 import android.util.AttributeSet;
