+++ app/src/main/java/flower/classifier/customview/OverlayView.java
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-package org.tensorflow.lite.examples.classification.customview;
+package flower.classifier.customview;
 
 import android.content.Context;
 import android.graphics.Canvas;
