+++ app/src/main/java/flower/classifier/summary.java
@@ -1,4 +1,4 @@
-package org.tensorflow.lite.examples.classification;
+package flower.classifier;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,6 +6,7 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.TextView;
+
 
 import java.util.ArrayList;
