+++ app/src/main/res/layout/tfe_ic_activity_camera.xml
@@ -31,7 +31,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="@android:color/black"
-            tools:context="org.tensorflow.lite.examples.classification.CameraActivity" >
+            tools:context="org.tensorflow.lite.examples.classification.flower.classifier.CameraActivity" >
 
             <Button
                 android:id="@+id/button8"
