+++ app/src/main/res/values/strings.xml
@@ -13,7 +13,7 @@
     <string name="tfe_ic_device" translatable="false">Device:</string>
     <string-array name="tfe_ic_devices" translatable="false">
         <item>CPU</item>
-        <!-- TODO: Add GPU  -->
+        <item>GPU</item>
 
     </string-array>
     <string name="new_picture">New picture</string>
